Correct css and add confirmation to delete users and treasures

+++ src/components/FormulariTreasure.vue
@@ -3,7 +3,7 @@
 
     <form id="form" @submit.prevent="sendTreasure">
       <div id="imgDiv" class="image-treasure">
-        <img id="preview" alt="preview"/>
+        <img id="preview" alt="preview" />
         <div class="treasure-image">
           <input id="image" type="file" @change="onFileUpload">
         </div>
@@ -20,12 +20,12 @@
         <div>
           <label for="latitude">Latitude </label>
           <input type="number" step="0.00001" id="latitude" min=-90 max=90 name="latitude" required
-                 v-model="treasure.latitude">
+            v-model="treasure.latitude">
         </div>
         <div>
           <label for="longitude">Longitude </label>
           <input type="number" step="0.00001" id="longitude" min=-180 max=180 name="longitude" required
-                 v-model="treasure.longitude">
+            v-model="treasure.longitude">
         </div>
         <div>
           <label for="location">Location </label>
@@ -52,16 +52,16 @@
           </select>
         </div>
         <div class="image-container">
-          <input v-if="id === ''" class="form-submit" type="submit" value="Add"/>
+          <input v-if="id === ''" class="form-submit" type="submit" value="Add" />
           <img class="buttons" v-if="id !== ''" @click.prevent="updateTreasure" src="../assets/icons/edit.png"
-               alt="edit-icon"/>
+            alt="edit-icon" />
           <img class="buttons" v-if="id !== ''" @click.prevent="deleteTreasure" src="../assets/icons/borrar.png"
-               alt="delete-icon"/>
+            alt="delete-icon" />
         </div>
       </div>
     </form>
     <MapComponent class="map" ref="mapa" v-bind:new="newTreasure" v-bind:id="this.$route.params.idTreasure"
-                  v-bind:size="size" v-on:emitCoords="receiveCoords"/>
+      v-bind:size="size" v-on:emitCoords="receiveCoords" />
   </div>
 </template>
 
@@ -72,7 +72,7 @@ import UserService from "@/services/user.service";
 
 export default {
   name: 'FormulariTreasure',
-  components: {MapComponent},
+  components: { MapComponent },
   props: {
     id: String
   },
@@ -131,13 +131,13 @@ export default {
       TreasureService.createNew(formData).then(() => {
         this.$router.push(`/map`)
       }).catch((error) => {
-            if (error.response.status === 401) {
-              localStorage.setItem('error', JSON.stringify("Unauthorized action"))
-              this.$router.push("/error")
-            }
-            localStorage.setItem('error', JSON.stringify("Error when creating new treasure"))
-            this.$router.push("/error")
-          }
+        if (error.response.status === 401) {
+          localStorage.setItem('error', JSON.stringify("Unauthorized action"))
+          this.$router.push("/error")
+        }
+        localStorage.setItem('error', JSON.stringify("Error when creating new treasure"))
+        this.$router.push("/error")
+      }
       );
 
     },
@@ -152,8 +152,8 @@ export default {
             bytes[i] = response.data.charCodeAt(i);
           }
 
-          const blob = new Blob(bytes, {type: response.headers["content-type"]})
-          this.FILE = new File([blob], this.treasure.image, {type: response.headers["content-type"]})
+          const blob = new Blob(bytes, { type: response.headers["content-type"] })
+          this.FILE = new File([blob], this.treasure.image, { type: response.headers["content-type"] })
 
 
           const formData = new FormData()
@@ -192,26 +192,27 @@ export default {
     },
 
     deleteTreasure() {
+      const confirmation = confirm("Are you sure you want to delete this treasure?");
+      if (confirmation) {
+        UserService.getCurrentUsername().then((response) => {
+          UserService.getByNickname(response.data).then((response) => {
 
-      UserService.getCurrentUsername().then((response) => {
-        UserService.getByNickname(response.data).then((response) => {
-
-          if (response.data.userRole === "Admin") {
-            TreasureService.delete(this.id).then((response) => {
-              alert(response.data)
-              this.$router.push("/management")
-            }).catch(() => {
-              localStorage.setItem('error', JSON.stringify("Error when deleting treasure"))
-              this.$router.push("/error")
-            })
-          }
+            if (response.data.userRole === "Admin") {
+              TreasureService.delete(this.id).then((response) => {
+                alert(response.data)
+                this.$router.push("/management")
+              }).catch(() => {
+                localStorage.setItem('error', JSON.stringify("Error when deleting treasure"))
+                this.$router.push("/error")
+              })
+            }
+          }).catch(() => {
+          })
         }).catch(() => {
+          localStorage.removeItem('token')
+          this.$router.push("/login")
         })
-      }).catch(() => {
-        localStorage.removeItem('token')
-        this.$router.push("/login")
-      })
-
+      }
     }
   },
   mounted() {
@@ -219,10 +220,10 @@ export default {
 
     if (this.id !== "") {
       TreasureService.getById(this.id).then((response) => {
-            this.treasure = response.data
-            this.getImage(response.data)
+        this.treasure = response.data
+        this.getImage(response.data)
 
-          }
+      }
       ).catch((error) => {
         localStorage.setItem('error', error.response.data)
         this.$router.push("/error")
@@ -281,7 +282,8 @@ export default {
 
 .map {
   z-index: 0;
-  margin-left: 10rem;;
+  margin-left: 10rem;
+  ;
 }
 
 .form-submit {
@@ -408,5 +410,6 @@ select {
   justify-content: end;
   margin-top: 20px;
   margin-left: auto;
-}</style>
+}
+</style>
